src/pyflask/requirements.txt
altgraph==0.17.2
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
flask-restx==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
pynmea2-1.18.0